move original logic_tree into version1 package;
