+++ nzshm_model/model.py
@@ -1,10 +1,10 @@
 import json
 import warnings
 from pathlib import Path
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Union
 
 from nzshm_model.psha_adapter import NrmlDocument, OpenquakeSimplePshaAdapter
-from nzshm_model.source_logic_tree.logic_tree import SourceLogicTree
+from nzshm_model.source_logic_tree import SourceLogicTree, SourceLogicTreeV1
 
 if TYPE_CHECKING:
     from nzshm_model.psha_adapter.openquake.logic_tree import LogicTree
@@ -26,12 +26,12 @@ class NshmModel:
         assert self._slt_json.exists()
         assert self._gmm_xml.exists()
 
-    def source_logic_tree(self) -> "SourceLogicTree":
+    def source_logic_tree(self) -> Union["SourceLogicTreeV1", "SourceLogicTree"]:
         with open(self._slt_json, 'r') as jsonfile:
             data = json.load(jsonfile)
             ltv = data.get("logic_tree_version")
             if ltv is None:  # original json is unversioned
-                return SourceLogicTree.from_dict(data)
+                return SourceLogicTreeV1.from_dict(data)
             raise ValueError("Unsupported logic_tree_version.")
 
     def source_logic_tree_nrml(self) -> "LogicTree":
