+++ nzshm_model/source_logic_tree/__init__.py
@@ -1,0 +1,4 @@
+from .branch_attribute import BranchAttributeSpec, BranchAttributeValue
+from .version1.logic_tree import SourceLogicTree as SourceLogicTreeV1
+from .version1.logic_tree import SourceLogicTreeCorrelation
+from .version2.logic_tree import SourceLogicTree
