+++ nzshm_model/source_logic_tree/version1/__init__.py
@@ -1,0 +1,1 @@
+from .logic_tree import Branch, FaultSystemLogicTree, SourceLogicTree, SourceLogicTreeCorrelation
