+++ nzshm_model/source_logic_tree/version1/logic_tree.py
@@ -11,37 +11,13 @@ from functools import reduce
 from itertools import product
 from math import isclose
 from operator import add, mul
-from typing import Any, Dict, Generator, Iterable, List, Type, Union
+from typing import Dict, Generator, Iterable, List, Type, Union
 
 import dacite
 
 from nzshm_model.psha_adapter import PshaAdapterInterface
 
-
-@dataclass
-class BranchAttributeSpec:
-    name: str
-    long_name: str
-    value_options: List[Any] = field(default_factory=list)
-
-
-@dataclass(frozen=True)
-class BranchAttributeValue:
-    name: str
-    long_name: str
-    value: Any = None
-
-    @staticmethod
-    def from_branch_attribute(ba: BranchAttributeSpec, value):
-        return BranchAttributeValue(ba.name, ba.long_name, value)
-
-    @staticmethod
-    def all_from_branch_attribute(ba: BranchAttributeSpec):
-        for opt in ba.value_options:
-            yield BranchAttributeValue(ba.name, ba.long_name, opt)
-
-    def __repr__(self):
-        return f"{self.name}{self.value}"
+from .. import BranchAttributeSpec, BranchAttributeValue
 
 
 @dataclass
